Profiling: Move to t-digest
t-digest is also a constant memory algorithm, or rather it has bounded
memory which I can pretty easily deduce at compile time. P2 quantiles
were nice, but the implementation used some magic I didn't understand
and had some wild bugs that I wasn't willing to fix.

New implementation is allows for any quantile to be obtained, but to not
have to change the entire world (only most of it) I opted to keep the
Tally abstraction (even though it just forwards it's calls directly to
TDigest).

Additionally, I no longer pass the entire Profiling stuct back over IPC,
just the Zonable part. Otherwise, since this implementation _is_ much
larger, but it's all on the stack so idrc.

Finally there's some funky stuff going on with how I'm passing around
profiling. I need to have it in the TestOpts struct at some point,
because `run` is one level up from `runChild`, so it's currently
initialized in `runAll` because some things in the logger also depend on
knowing if Profiling is null or not. This is ugly and annoying but I
don't immediately know a good solution. Will have to fix this later.