tests: Flatten file structure
I didn't like the mimalloc-bench namespace so this goes most the way to
remove it. Besides that tests are now loosely organized into their
function, currently being correctness, micro(bechmark) and non
catagarized ones (for now) being mimalloc ported tests.

In the future I want to add security tests, ones that don't neccesarily
aim for correctness, but aim to fuck with the allocator. To do that
however I'll have to reach into the runner to make it understand that
these tests may fail, and that that is OK, but should be reported.

Also the currently unused tests in root as well as the uncategorized
tests in mimalloc-bench.zig need to be categorized and or removed at
some point